LayoutTests/platform/ios-simulator/fast/css/style-scoped/style-scoped-change-scoped-in-shadow-expected.txt

CONSOLE MESSAGE: line 3: HierarchyRequestError: DOM Exception 3: A Node was inserted somewhere it doesn't belong.
PASS window.internals is defined.
PASS successfullyParsed is true

TEST COMPLETE
test a scoped style in shadow subtree.
